requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.0
panda==0.3.1
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
threadpoolctl==3.6.0
urllib3==2.4.0
Werkzeug==3.1.3
